fix: avatar src loading improvements
